+++ pymc3/distributions/__init__.py
@@ -26,6 +26,7 @@ from .continuous import SkewNormal
 from .continuous import Triangular
 from .continuous import Gumbel
 from .continuous import Logistic
+from .continuous import LogitNormal
 from .continuous import Interpolated
 
 from .discrete import Binomial
@@ -145,6 +146,7 @@ __all__ = ['Uniform',
            'DiscreteWeibull',
            'Gumbel',
            'Logistic',
+           'LogitNormal',
            'Interpolated',
            'Bound',
            ]